app/cfo_voice/requirements.txt
Flask==3.0.3
flask-cors>=4.0.0
google-generativeai>=0.6.0
elevenlabs>=1.6
requests>=2.32
pydub>=0.25.1
SpeechRecognition>=3.10
python-dotenv>=1.0.1
gunicorn>=21.2
